Customer .h and Customer.cpp (#12)
* Customer .h and Customer.cpp

These are the customer files sorry for the delay team my bad

* Added Vehicle.cpp

* Final update for today

DisplayCustomer() method is finished now

* Changes to Customer.h

I made a mistake and made it in c and not c++

* main()

Hi, please add or voice your suggestions on what CalculateSales()  will effect the variables salesDay, salesWeek, and salesMonth. Set them to a default value for example. Also add the 2 calls of readfile() if you know the names of the both of the files we are reading. Thank you

---------

Co-authored-by: Marco+Marshall <mdrook@student.cccs.edu>
Co-authored-by: Jack <sparkdragon586@gmail.com>