Added Search Vehicle and struct (#11)
* Basic Search Code

Made a struct for sorting. Prompted the user. Added up the amount of each type of car. Called the display function. Added getSold and setSold to Vehicle.h.

* Fixed Bugs

Fixed a few bugs as well as made validation better.
Please make the function in the Vehicle.h file or our code wont work. Also please add the vector in main.

* Comments

Added the IPO for this function

* Deleted

Deleted non relevant files to project

* Added string

Added a string to the display vehicle function. Added a string that will be passed in to the display vehicle function.

---------

Co-authored-by: Jack <sparkdragon586@gmail.com>